change: グラフ調整

+++ server/views/sub/js/index.js
@@ -47,11 +47,12 @@
 
     $(() => {
         $(button).on('click', () => {
-            chart.data.labels.splice(0, 1);
-            chart.data.datasets[0].data.splice(0, 1);
-            chart.data.labels.push('新しいの');
-            chart.data.datasets[0].data.push(100);
-            chart.update();
+            // chart.data.labels.splice(0, 1);
+            // chart.data.datasets[0].data.splice(0, 1);
+            // chart.data.labels.push('新しいの');
+            // chart.data.datasets[0].data.push(100);
+            // chart.update();
+            console.log(chart.data);
         });
 
         socket.on('add date', (date) => {
@@ -76,26 +77,46 @@
         });
 
         socket.on('graph update', (datas) => {
-            chart.data.labels.splice(0, 1);
-            // chart.data.datasets[0].data.splice(0, 1);
-            $.graph.delete();
-            console.log(datas);
+            // 一番古いデータ削除
+
+            if(chart.data.labels.length >= 10){
+                chart.data.labels.splice(0, 1);
+                $.graph.delete();
+            }
+
+            console.log(chart.data.datasets);
             datas.forEach((data, index) => {
-                if (chart.data.datasets[index] == undifinde) {
-                    chart.data.datasets[index].push({ backgroundColor: $.graph.randCode(), label: family[Math.floor(Math.random() * 1)], data: [] });
-                    const length = chart.data.labels.length();
-                    console.log(length);
-                    chart.data.datasets[index].data[length] = data;
+                // datasetsの中身か存在しなかった時の処理
+                if (chart.data.datasets[index] === undefined) {
+                    const newDatasets = {
+                        backgroundColor: $.graph.randCode(),
+                        label: family[(index + 1) % 2],
+                        data: [],
+                    };
+                    // const length = chart.data.labels.length;
+                    // console.log(length);
+                    for (let i = 0; i <= chart.data.labels.length - 1; i++) {
+                        newDatasets.data.push(0);
+                    }
+                    newDatasets.data.push(data);
+                    chart.data.datasets.push(newDatasets);
                 } else {
                     chart.data.datasets[index].data.push(data);
                 }
             });
 
-            chart.data.labels.push('新しいの');
-            $.graph.add();
-            chart.data.datasets[0].data.push(data);
+            if (datas.length < chart.data.datasets.length) {
+                for (let len = datas.length; len < chart.data.datasets.length; len++) {
+                    chart.data.datasets[len].data.push(0);
+                }
+            }
+            const time = new Date();
+
+            chart.data.labels.push(`${time.getHours()}:${time.getMinutes()}`);
+            // $.graph.add();
+            // chart.data.datasets[0].data.push(data);
             chart.update();
-            console.log(chart.data.datasets);
+            // console.log(chart.data.datasets);
         });
     });
 
@@ -110,7 +131,7 @@
                 dataset.data.shift();
             });
         },
-        randCode: () => `rgb(${Math.random() * 101},${Math.random() * 101},${Math.random() * 101})`,
+        randCode: () => `rgb(${Math.floor(Math.random() * 255)},${Math.floor(Math.random() * 255)},${Math.floor(Math.random() * 255)})`,
         addDataset: () => {
             const length = chart.data.labels.length();
         },
